requirements.txt
tensorflow
tensorflow-datasets
pillow
matplotlib
scikit-learn
kaggle
numpy
pandas
mlflow
gradio